pdf-parser-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pdfplumber==0.11.0
python-multipart==0.0.9
python-dotenv==1.0.1
openai>=1.12.0
httpx>=0.27.0
